Modularize twisted edwards extended templates
