+++ curves/src/templates/twisted_edwards_extended/projective.rs
@@ -1,0 +1,391 @@
+// Copyright (C) 2019-2021 Aleo Systems Inc.
+// This file is part of the snarkVM library.
+
+// The snarkVM library is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+
+// The snarkVM library is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+// GNU General Public License for more details.
+
+// You should have received a copy of the GNU General Public License
+// along with the snarkVM library. If not, see <https://www.gnu.org/licenses/>.
+
+use crate::{
+    impl_edwards_curve_serializer,
+    templates::twisted_edwards_extended::GroupAffine,
+    traits::{
+        AffineCurve,
+        Group,
+        MontgomeryModelParameters as MontgomeryParameters,
+        ProjectiveCurve,
+        TEModelParameters as Parameters,
+    },
+};
+use snarkvm_fields::{impl_additive_ops_from_ref, Field, One, PrimeField, SquareRootField, Zero};
+use snarkvm_utilities::{
+    bititerator::BitIteratorBE,
+    bytes::{FromBytes, ToBytes},
+    rand::UniformRand,
+    serialize::*,
+};
+
+use rand::{
+    distributions::{Distribution, Standard},
+    Rng,
+};
+use serde::{Deserialize, Serialize};
+use std::{
+    fmt::{Display, Formatter, Result as FmtResult},
+    io::{Read, Result as IoResult, Write},
+    ops::{Add, AddAssign, Mul, MulAssign, Neg, Sub, SubAssign},
+};
+
+#[derive(Derivative)]
+#[derivative(
+    Copy(bound = "P: Parameters"),
+    Clone(bound = "P: Parameters"),
+    Eq(bound = "P: Parameters"),
+    Debug(bound = "P: Parameters"),
+    Hash(bound = "P: Parameters")
+)]
+pub struct GroupProjective<P: Parameters> {
+    pub x: P::BaseField,
+    pub y: P::BaseField,
+    pub t: P::BaseField,
+    pub z: P::BaseField,
+}
+
+impl<P: Parameters> GroupProjective<P> {
+    pub fn new(x: P::BaseField, y: P::BaseField, t: P::BaseField, z: P::BaseField) -> Self {
+        Self { x, y, t, z }
+    }
+}
+
+impl<P: Parameters> Display for GroupProjective<P> {
+    fn fmt(&self, f: &mut Formatter<'_>) -> FmtResult {
+        write!(f, "{}", self.into_affine())
+    }
+}
+
+impl<P: Parameters> PartialEq for GroupProjective<P> {
+    fn eq(&self, other: &Self) -> bool {
+        if self.is_zero() {
+            return other.is_zero();
+        }
+
+        if other.is_zero() {
+            return false;
+        }
+
+        // x1/z1 == x2/z2  <==> x1 * z2 == x2 * z1
+        (self.x * other.z) == (other.x * self.z) && (self.y * other.z) == (other.y * self.z)
+    }
+}
+
+impl<P: Parameters> Distribution<GroupProjective<P>> for Standard {
+    #[inline]
+    fn sample<R: Rng + ?Sized>(&self, rng: &mut R) -> GroupProjective<P> {
+        loop {
+            let x = P::BaseField::rand(rng);
+            let greatest = rng.gen();
+
+            if let Some(p) = GroupAffine::from_x_coordinate(x, greatest) {
+                return p.scale_by_cofactor();
+            }
+        }
+    }
+}
+
+impl<P: Parameters> ToBytes for GroupProjective<P> {
+    #[inline]
+    fn write<W: Write>(&self, mut writer: W) -> IoResult<()> {
+        self.x.write(&mut writer)?;
+        self.y.write(&mut writer)?;
+        self.t.write(&mut writer)?;
+        self.z.write(writer)
+    }
+}
+
+impl<P: Parameters> FromBytes for GroupProjective<P> {
+    #[inline]
+    fn read<R: Read>(mut reader: R) -> IoResult<Self> {
+        let x = P::BaseField::read(&mut reader)?;
+        let y = P::BaseField::read(&mut reader)?;
+        let t = P::BaseField::read(&mut reader)?;
+        let z = P::BaseField::read(reader)?;
+        Ok(Self::new(x, y, t, z))
+    }
+}
+
+impl<P: Parameters> Default for GroupProjective<P> {
+    #[inline]
+    fn default() -> Self {
+        Self::zero()
+    }
+}
+
+impl<P: Parameters> Zero for GroupProjective<P> {
+    fn zero() -> Self {
+        Self::new(
+            P::BaseField::zero(),
+            P::BaseField::one(),
+            P::BaseField::zero(),
+            P::BaseField::one(),
+        )
+    }
+
+    fn is_zero(&self) -> bool {
+        self.x.is_zero() && self.y == self.z && !self.y.is_zero() && self.t.is_zero()
+    }
+}
+
+impl<P: Parameters> ProjectiveCurve for GroupProjective<P> {
+    type Affine = GroupAffine<P>;
+    type BaseField = P::BaseField;
+
+    fn prime_subgroup_generator() -> Self {
+        GroupAffine::prime_subgroup_generator().into()
+    }
+
+    fn is_normalized(&self) -> bool {
+        self.z.is_one()
+    }
+
+    fn batch_normalization(v: &mut [Self]) {
+        // Montgomery’s Trick and Fast Implementation of Masked AES
+        // Genelle, Prouff and Quisquater
+        // Section 3.2
+
+        // First pass: compute [a, ab, abc, ...]
+        let mut prod = Vec::with_capacity(v.len());
+        let mut tmp = P::BaseField::one();
+        for g in v
+            .iter_mut()
+            // Ignore normalized elements
+            .filter(|g| !g.is_normalized())
+        {
+            tmp.mul_assign(&g.z);
+            prod.push(tmp);
+        }
+
+        // Invert `tmp`.
+        tmp = tmp.inverse().unwrap(); // Guaranteed to be nonzero.
+
+        // Second pass: iterate backwards to compute inverses
+        for (g, s) in v
+            .iter_mut()
+            // Backwards
+            .rev()
+            // Ignore normalized elements
+            .filter(|g| !g.is_normalized())
+            // Backwards, skip last element, fill in one for last term.
+            .zip(
+                prod.into_iter()
+                    .rev()
+                    .skip(1)
+                    .chain(Some(P::BaseField::one())),
+            )
+        {
+            // tmp := tmp * g.z; g.z := tmp * s = 1/z
+            let newtmp = tmp * g.z;
+            g.z = tmp * s;
+            tmp = newtmp;
+        }
+
+        // Perform affine transformations
+        for g in v.iter_mut().filter(|g| !g.is_normalized()) {
+            g.x *= &g.z; // x/z
+            g.y *= &g.z;
+            g.t *= &g.z;
+            g.z = P::BaseField::one(); // z = 1
+        }
+    }
+
+    #[allow(clippy::many_single_char_names)]
+    fn add_assign_mixed(&mut self, other: &Self::Affine) {
+        // A = X1*X2
+        let a = self.x * other.x;
+        // B = Y1*Y2
+        let b = self.y * other.y;
+        // C = T1*d*T2
+        let c = P::COEFF_D * self.t * other.x * other.y;
+        // D = Z1
+        let d = self.z;
+        // E = (X1+Y1)*(X2+Y2)-A-B
+        let e = (self.x + self.y) * (other.x + other.y) - a - b;
+        // F = D-C
+        let f = d - c;
+        // G = D+C
+        let g = d + c;
+        // H = B-a*A
+        let h = b - P::mul_by_a(&a);
+        // X3 = E*F
+        self.x = e * f;
+        // Y3 = G*H
+        self.y = g * h;
+        // T3 = E*H
+        self.t = e * h;
+        // Z3 = F*G
+        self.z = f * g;
+    }
+
+    fn into_affine(&self) -> GroupAffine<P> {
+        (*self).into()
+    }
+
+    fn recommended_wnaf_for_scalar(scalar: <Self::ScalarField as PrimeField>::BigInteger) -> usize {
+        P::empirical_recommended_wnaf_for_scalar(scalar)
+    }
+
+    fn recommended_wnaf_for_num_scalars(num_scalars: usize) -> usize {
+        P::empirical_recommended_wnaf_for_num_scalars(num_scalars)
+    }
+}
+
+impl<P: Parameters> Group for GroupProjective<P> {
+    type ScalarField = P::ScalarField;
+
+    #[inline]
+    #[must_use]
+    fn double(&self) -> Self {
+        let mut tmp = *self;
+        tmp += self;
+        tmp
+    }
+
+    #[inline]
+    fn double_in_place(&mut self) {
+        let tmp = *self;
+        *self = tmp.double();
+    }
+}
+
+impl<P: Parameters> Neg for GroupProjective<P> {
+    type Output = Self;
+
+    fn neg(mut self) -> Self {
+        self.x = -self.x;
+        self.t = -self.t;
+        self
+    }
+}
+
+impl_additive_ops_from_ref!(GroupProjective, Parameters);
+
+impl<'a, P: Parameters> Add<&'a Self> for GroupProjective<P> {
+    type Output = Self;
+
+    fn add(self, other: &'a Self) -> Self {
+        let mut copy = self;
+        copy += other;
+        copy
+    }
+}
+
+impl<'a, P: Parameters> AddAssign<&'a Self> for GroupProjective<P> {
+    #[allow(clippy::many_single_char_names)]
+    #[allow(clippy::suspicious_op_assign_impl)]
+    fn add_assign(&mut self, other: &'a Self) {
+        // See "Twisted Edwards Curves Revisited"
+        // Huseyin Hisil, Kenneth Koon-Ho Wong, Gary Carter, and Ed Dawson
+        // 3.1 Unified Addition in E^e
+
+        // A = x1 * x2
+        let a = self.x * other.x;
+
+        // B = y1 * y2
+        let b = self.y * other.y;
+
+        // C = d * t1 * t2
+        let c = P::COEFF_D * self.t * other.t;
+
+        // D = z1 * z2
+        let d = self.z * other.z;
+
+        // H = B - aA
+        let h = b - P::mul_by_a(&a);
+
+        // E = (x1 + y1) * (x2 + y2) - A - B
+        let e = (self.x + self.y) * (other.x + other.y) - a - b;
+
+        // F = D - C
+        let f = d - c;
+
+        // G = D + C
+        let g = d + c;
+
+        // x3 = E * F
+        self.x = e * f;
+
+        // y3 = G * H
+        self.y = g * h;
+
+        // t3 = E * H
+        self.t = e * h;
+
+        // z3 = F * G
+        self.z = f * g;
+    }
+}
+
+impl<'a, P: Parameters> Sub<&'a Self> for GroupProjective<P> {
+    type Output = Self;
+
+    fn sub(self, other: &'a Self) -> Self {
+        let mut copy = self;
+        copy -= other;
+        copy
+    }
+}
+
+impl<'a, P: Parameters> SubAssign<&'a Self> for GroupProjective<P> {
+    fn sub_assign(&mut self, other: &'a Self) {
+        *self += &(-(*other));
+    }
+}
+
+impl<P: Parameters> Mul<P::ScalarField> for GroupProjective<P> {
+    type Output = Self;
+
+    /// Performs scalar multiplication of this element.
+    #[allow(clippy::suspicious_arithmetic_impl)]
+    #[inline]
+    fn mul(self, other: P::ScalarField) -> Self {
+        let mut res = Self::zero();
+
+        let mut found_one = false;
+
+        for i in BitIteratorBE::new(other.into_repr()) {
+            if found_one {
+                res.double_in_place();
+            } else {
+                found_one = i;
+            }
+
+            if i {
+                res += self;
+            }
+        }
+
+        res
+    }
+}
+
+impl<P: Parameters> MulAssign<P::ScalarField> for GroupProjective<P> {
+    /// Performs scalar multiplication of this element.
+    fn mul_assign(&mut self, other: P::ScalarField) {
+        *self = *self * other
+    }
+}
+
+// The affine point (X, Y) is represented in the Extended Projective coordinates
+// with Z = 1.
+impl<P: Parameters> From<GroupAffine<P>> for GroupProjective<P> {
+    fn from(p: GroupAffine<P>) -> GroupProjective<P> {
+        Self::new(p.x, p.y, p.x * p.y, P::BaseField::one())
+    }
+}
